move ligature_carets out of googlefonts directory
