+++ Lib/fontbakery/checks/ligature_carets.py
